UI-XX: Set host header on proxy to request host
If downstream uses host based routing (like orderweb to switch country
based on tld) then we need the original host to be passed through
without being modified. Otherwise the host will be changed to the
internal nginx host. Prior to this change the host would have been set
to be `app`.