Allow client_body_buffer_size to be overriden.
This change allows `client_body_buffer_size` to be overriden by users of
the sidecar. It defaults the environment variable to 8k, which is what
nginx sets it to on x86-64.

We're regularly seeing larger requests get written to disk on some of
our services. This will allow us to configure a larger buffer size to
avoid this.

In the future, we may want to change the default 8k size to something
larger.