Add support for KEEPALIVE_TIMEOUT env variable
The env variable can be used to override the default keepalive_timeout, which was historically set to 20s in this image.

The reason for making keepalive_timeout configurable is to align with other timeouts along the path of an HTTP request.
A common topology in our apps is ALB -> Nginx -> Rails. If the keep-alive timeout for Nginx and Rails is lower than the timeout for the ALB, the latter will report some HTTP 502 errors whenever one of its connections is killed by the downstream systems because of a timeout.